refactor functions to own file
